requirements.txt
RPi.GPIO==0.7.1
PyYAML==6.0.1
matplotlib==3.8.2
